fix set len

+++ decode/redisObject.go
@@ -7,10 +7,10 @@ package main
  */
 type RedisObject struct {
 	objType int
-	objLen  int
+	objLen  int64
 	objVal  interface{}
 }
 
-func NewRedisObject(objType int, objLen int, objVal interface{}) *RedisObject {
+func NewRedisObject(objType int, objLen int64, objVal interface{}) *RedisObject {
 	return &RedisObject{objType, objLen, objVal}
 }
